+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./digitalPlanningSchema";
 export * from "./encryption";
 export * from "./govPayMetadata";
+export * from "./projectTypes";
